hw/core: Fix the order of the roms list. (#219)
The original logic for comparing the order involves directly comparing pointers,
which may lead to an unintended order of the roms list.
For exmple, using a generic loader to load a file into memory in Windows may
failed because the content is overwritten by other roms.

We change the logic as follows:
1. If pointers of AddressSpace are different, return greater.
2. If pointers of AddressSpace are the same, return greater if the address
   is greater.

Signed-off-by: Yu-Ming Chang <yumin686@andestech.com>

Reviewed-on: https://gitea.andestech.com/RD-SW/qemu/pulls/219
Reviewed-by: Alvin Che-Chia Chang <alvinga@andestech.com>
Co-authored-by: Yu-Ming Chang <yumin686@andestech.com>
Co-committed-by: Yu-Ming Chang <yumin686@andestech.com>