hw/riscv: Support boot ROM from memory-backend-file (#227)
This allows the AE350 machine to optionally use an external
memory-backend-file object as the boot ROM (MROM) region.

If the user specified a backend via -object (e.g.,
-object memory-backend-file,id=flash-backend,mem-path=boot.img,size=...),
use it as the boot ROM. This enables preloaded flash content or persistent storage.

Otherwise, the machine falls back to creating a local ROM region
and loads firmware into it as before.

This enables persistent flash simulation, e.g., for secure boot
development or testing firmware upgrades.

Signed-off-by: Robin Chun-An Lu <robin610@andestech.com>

Reviewed-on: https://gitea.andestech.com/RD-SW/qemu/pulls/227
Reviewed-by: Yuming Yu-Ming Chang <yumin686@andestech.com>
Co-authored-by: robin610 <robin610@andestech.com>
Co-committed-by: robin610 <robin610@andestech.com>