hw/timer: Check Andes PLMT timecmp values to avoid QEMU timer overflow. (#238)
When we set a timecmp value, it will be converted to a timer_ns value
based on its frequency to set a QEMU timer. The converted timer_ns value
could be greater than INT64_MAX, indicating that the expire time of QEMU timer
could overflow.
We can calculate the INT64_MAX timer_ns back to the max timecmp value. If
a timecmp value exceeds the max timecmp value, an overflow will occur,
and we can limit the expire time of the QEMU timer to INT64_MAX.

Signed-off-by: Yu-Ming Chang <yumin686@andestech.com>
Reviewed-on: https://gitea.andestech.com/RD-SW/qemu/pulls/238
Reviewed-by: Ethan Zhan-Yang Chen <ethan84@andestech.com>
Co-authored-by: Yu-Ming Chang <yumin686@andestech.com>
Co-committed-by: Yu-Ming Chang <yumin686@andestech.com>