semihosting: Fix return value of SYS_FLEN operation. (#242)
Bug 35603

If the SYS_FLEN request is handled by the GDB, QEMU sends an 'Ffstat'
packet to GDB and receives a return value indicating either success or
an error. This return value is a status code and should not be compared
with the file size.
The return value of SYS_FLEN request should be the file size, not the
status code itself.

Signed-off-by: Yu-Ming Chang <yumin686@andestech.com>
Reviewed-on: https://gitea.andestech.com/RD-SW/qemu/pulls/242
Reviewed-by: Foxes Cheng-Hung Hung <foxes687@andestech.com>
Co-authored-by: Yu-Ming Chang <yumin686@andestech.com>
Co-committed-by: Yu-Ming Chang <yumin686@andestech.com>