target/riscv: Allow CONTEXT, CS bit to be enabled (#243)
Bug 35818

Currently, we cannot enable the CONTEXT and CS bits in the stateen register.
These bits are required to be writable according to the AX66 specification and
should not affect other Andes cores.

Although we are not refining the write mask at this time, we may consider doing
so in the future if there are Andes cores that fully comply with the RISC-V
Debug Specification v1.0.

Signed-off-by: Leo Yu-Chi Liang <ycliang@andestech.com>

Reviewed-on: https://gitea.andestech.com/RD-SW/qemu/pulls/243
Co-authored-by: Leo Yu-Chi Liang <ycliang@andestech.com>
Co-committed-by: Leo Yu-Chi Liang <ycliang@andestech.com>