hw/sd: Fix issue where SD card on atfsdc010 remains in busy state. (#246)
Before sending a command to the SD card, the device driver checks the
DATA0_STATUS bit to ensure the card is idle. However, the controller
initializes this bit incorrectly, causing the SD card to appear busy
and preventing commands from being issued.

This patch initializes DATA0_STATUS to the idle state so that the
device driver can pass its check and proceed with command execution.

Signed-off-by: Yu-Ming Chang <yumin686@andestech.com>

Reviewed-on: https://gitea.andestech.com/RD-SW/qemu/pulls/246
Reviewed-by: CL Chin-Long Wang <cl634@andestech.com>
Reviewed-by: randolph <randolph@andestech.com>
Reviewed-by: Foxes Cheng-Hung Hung <foxes687@andestech.com>
Co-authored-by: Yu-Ming Chang <yumin686@andestech.com>
Co-committed-by: Yu-Ming Chang <yumin686@andestech.com>